Differential Expression Analysis Using EdgeR

Perform DEA by EdgeR on normalized dataset or dataset with adjusting factor.

DE.edgeR(RC, groups, Pval = 0.01, normalized = TRUE, adjust = NULL)

Arguments

RC
data in the format of data frame or matrix, with columns for samples and raws for genes.

groups
vector of characters indicating the group for each sample.

Pval
cut-off point for p-values to identify DE genes.

normalized
whether the dataset is normalized (by scaling method)

adjust
the adjusting factor if the dataset itself is not normalized

Value

list, containing id.list (names of DE genes), p.val, and log2.FC.

References

https://ucdavis-bioinformatics-training.github.io/2018-June-RNA-Seq-Workshop/thursday/DE.html

Examples

edgeR.benchmark <- DE.edgeR(data.benchmark, data.group)