Statistics for DEA Results Based Golden Standards

Computes the agreement of differential expression (DE) wihtin one data set with an assumed true DE status usually given based on a gold standard. The following statistics are computed based of the differentially expressed markers id.list compared to the list truly differentially expressed markers truth:

False Discovery Rate (FDR)

False Negative Rate (FNR)

True Positive Rate (TPR)

False Positive Rate (FPR)

DE.statistics(markers, id.list, truth, selected.marker = NULL)

Arguments

markers
Vector of all markers considered in the analysis.

id.list
Vector of markers that are differentially expressed in a data set --- typically in the data set data.test. Differential expression can be computed via DE.edgeR or DE.voom.

truth
Vector of genes that are (assumedly) truly differential expressed, e.g. basen on a gold standard --- typically the data set data.benchmark.

selected.marker
optional Vector of a subset of markers. If given, the analysis will be limited to the given subset. Leave NULL if all markers are considered for the analysis.

Value

A list of:

TPR

True positive Rate

FPR

False Positive Rate

FDR

False Discovery Rate

FNR

False Negative Rate

Details

Both, id.list and truth are a subset of markers from markers.

Examples

DE.bench <- DE.voom(data.benchmark, data.group)
DE.test <- DE.voom(data.test, data.group)
stats <- DE.statistics(rownames(data.benchmark), DE.test$id.list, DE.bench$id.list)
print(stats)
#> $TPR
#> [1] 0.4915254
#> 
#> $FPR
#> [1] 0.04209446
#> 
#> $FDR
#> [1] 0.5857143
#> 
#> $FNR
#> [1] 0.5084746
#>