MiRNA Sequencing Test Data

Myxofibrosarcoma (MXF) and pleomorphic malignant fibrous histiocytoma (PMFH) are the two most common and aggressive subtypes of genetically complex soft tissue sarcoma. MiRNAs for the same 54 tumors used for the benchmark data were re-sequenced using neither uniform handling nor balanced library assignment. In this study these samples were sequenced in the order of sample collection and processed in multiple runs.

data.test

Format

A dataframe with 1033 rows and 54 columns. Here are the examples of the column and row naming rule:

MXF2516

One sample belonging to MXF sample ID MXF2516.

hsa-let-7a-2*

Gene ID.