Concordance At The Top Plot

Function for generating concordance at the top plot, which compares concordance of the p-values obtained from benchmark data without normalization and normalized test data.

fig.CAT(
  MethodsCompare,
  benchmark = data.benchmark,
  test = data.test,
  group = data.group,
  DE.method = "DE.voom",
  QN_filter = FALSE,
  Pval = 0.01,
  MethodNew,
  pvalues,
  Methods_visual = c(MethodsCompare, MethodNew)
)

Arguments

MethodsCompare
the vector of methods that researchers would like to compare with, and selected from norm.none, norm.TMM, norm.TC, norm.UQ, norm.med, norm.DESeq, norm.PoissonSeq, norm.QN, norm.SVA, and norm.RUV.

benchmark
benchmark data with the same format of data.benchmark, which are used for benchmark p-values, and is set to use data.benchmark by default

test
test data with the same format of data.test, which are used for test p-values, and is set to use data.test by default

DE.method
the DEA method selected from DE.voom and DE.edgeR, and default to be DE.voom.

QN_filter
whether the filtering is performed if norm.QN is used.

Pval
p-value cut-off point for identifying DE genes, default to be 0.01.

MethodNew
the name of the new method which will be presented in the legend.

pvalues
a vector of p values computed from DE analysis using the new method normalized test data.

Methods_visual
a vector of methods' names shown in the figure.

groups
vector of characters indicating the group for each sample.

Value

figure of concordance for comparison

Examples

t <-  runif(1033)
names(t) <-  rownames(data.test)
fig.CAT(MethodsCompare = c("norm.none", "norm.TMM", "norm.SVA", "norm.TC", "norm.RUVr"), MethodNew = "Example", pvalues = t)

#> Number of significant surrogate variables is:  1 
#> Iteration (out of 5 ):1  2  3  4  5  

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'annotatedDataFrameFrom': could not find function "counts"