Venn diagram for p-values

Venn diagram is used to identify the performance of different normalization methods based on intersection of differential expressed genes.

fig.venn(truth.DEA, DEA.res, Pvalue)

Arguments

truth.DEA
differential expression analysis results from the benchmark (gold standard) obtained from DE.voom, DE.edge, or any method from the users by storing the results as same as DE methods in the package (including DE genes, p-values and log2 fold changes)

DEA.res
p-values as a result from prior differential expression analysis, e.g. using DE.voom or DE.edge.

Pvalue
Cut-off point for p-values for identifying significant differential expression.

Value

A Venn diagram

Examples

benchmark.voom <- DE.voom(RC = data.benchmark, groups = data.group, P = 0.01)
test.voom <- DE.voom(RC = data.test, groups = data.group, P = 0.01)
fig.venn(benchmark.voom$p.val, test.voom$p.val, 0.01)
#> Error: $ operator is invalid for atomic vectors