Normalization By Trimmed Mean of M-values (TMM)

Normalize the dataset using TMM, and return the normalized dataset with scaling factor scaled by 1e6.

norm.TMM(raw, groups)

Arguments

raw
raw count data in the format of data frame or matrix, with columns for samples and rows for genes.

groups
vector of characters indicating the group for each sample.

Value

list, containing dat.normed (normalized dataset) and scaling.factor (scaling factor) for each sample.

References

edgeR User Guide

Examples

test.TMM <-  norm.TMM(data.test, data.group)