Statistics for DEA Results Based Golden Standards

Computing the true positive rate, false postive rate, false discovery rate, and false negative rate based on given golden standards.

pip.statistics(
  raw,
  groups,
  norm.method,
  RUV_method = NULL,
  QN_filter = FALSE,
  DE.method = "DE.voom",
  Pval = 0.01,
  truth,
  marker_selection = FALSE,
  selected_marker = NULL
)

Arguments

raw
raw count data in the format of data frame or matrix, with columns for samples and raws for genes.

groups
vector of characters indicating the group for each sample (only 2 groups allowed).

norm.method
the method for normalization selected from norm.none, norm.TMM, norm.TC, norm.UQ, norm.med, norm.DESeq, norm.PoissonSeq, norm.QN, norm.SVA, and norm.RUV.

RUV_method
the exact RUV method used from RUVg, RUVr and RUVs if method = norm.RUV

QN_filter
whether the filtering is performed if method = norm.QN.

DE.method
the method for differential expression analysis from DE.voom and DE.edgeR, default to be DE.voom.

Pval
p-value for identifying DE genes, default to be 0.01

truth
vector of genes that are truly differential expressed

marker_selection
whether selecting a subset of genes/markers for this analysis

selected_marker
vector of a subset of genes/markers for this analysis

Value

a list of values about TPR, FPR, FDR, FNR

Examples

truthgene <- DE.voom(data.benchmark, data.group)$id.list
stat <- pip.statistics(data.test, data.group, "norm.TMM", truth = truthgene)