Full normalization assessment for given normalized test data

Full normalization assessment for given normalized test data

precision.seq(
  norm.counts,
  adjust.factors,
  method.name,
  DE.method = "DE.voom",
  Pval = 0.01
)

Arguments

norm.counts
Normalized counts of the test data for a user-provided normalization method.

adjust.factors
optional Adjustment factors for the normalization of the test data for a user-provided normalization method.

method.name
optional Name of the normalization method. Used for naming of data frames and plot legends.

DE.method
Method for computing differential expression statuses. Available: "DE.voom" and "DE.edgeR".

Pval
P-value cutoff for differential expression.

DEA
result for the data under study

truth.DEA
DEA result for the assumed truth (the gold standard)

title
Plot title

maxrank
Optionally specify the maximum size of top-ranked items that you want to plot.

subset
vector of a subset of genes/markers for this analysis

Value

List of assessment statistics and plots (ggplot objects):

data.norm

List of Normalized test data containing the test data for each package-supplied normalization methods and the normalized test data provided by the user.

data.DE

List of results of the differential expression analysis using the differential expression method DE.method for each normalization method using the test data.

benchmark.DE

Results of the differential expression analysis using the method DE.method for the benchmark data.

DE.stats

List of statistics (true positive rate, false positive rate, false discovery rate, and false negative rate) based on the comparison of differential expression statuses between the normalized test data and the unnormalized gold standard (the benchmark data).

p.volcano

List of volcano plots from the differential expression analysis for each normalization method.

p.RLE

List of Relative Log Expression plots for the normalized and un-normalized test data and the un-normalized benchmark data.

p.CAT

Concordance At The Top plot containing all normalization methods.

p.FNR_FDR

Plot of the False Discovery Rate over the False Negative Rate of the concordance of the differential expression statuses of the normalized test data with the ones of the gold standard (the benchmark) for each normalization method under study.

p.dendrogram

Dendrogram from clustering p-values from the differential expression analysis for each normalization method.

p.venn

List of venn diagrams of the correspondance of differential expression statuses between the normalized test data and the gold standard (the benchmark).

a list of values about TPR, FPR, FDR, FNR

Examples