Differential Expression Analysis Using EdgeR

Perform DEA by EdgeR on normalized dataset or dataset with adjusting factor. The normalized data can be provided as normalized counts or by adjusting factor for the original count data.

DE.edgeR(RC, groups, Pval = 0.01, normalized = TRUE, adjust = NULL)

Arguments

RC
data in the format of data frame or matrix, with columns for samples and raws for genes.

groups
vector of characters indicating the group for each sample.

Pval
cut-off point for p-values to identify DE genes.

normalized
Logical, whether the data is provided as normalized counts. If set to FALSE, adjustment factors must be provided using the adjust parameter.

adjust
Adjusting factors for normalizing the count data. Must be provided if the data is not normalized beforehand as indicated by the normalized parameter.

Value

list, containing id.list (names of DE genes), p.val, and log2.FC.

References

https://ucdavis-bioinformatics-training.github.io/2018-June-RNA-Seq-Workshop/thursday/DE.html

Examples

edgeR.benchmark <- DE.edgeR(data.benchmark, data.group)
str(edgeR.benchmark)
#> List of 3
#>  $ id.list: chr [1:71] "hsa-miR-183" "hsa-miR-182" "hsa-miR-129-2-3p" "hsa-miR-96" ...
#>  $ p.val  : num [1:1033] 2.53e-10 4.34e-10 1.42e-07 2.58e-07 7.56e-07 ...
#>  $ log2.FC: num [1:1033] 3.9 3.45 4.28 3.08 1.82 ...